+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 ### Changed
+- [Issue 370](https://github.com/aza547/wow-recorder/issues/370) - Show more than the most recent video on home page.
 ### Fixed
 
 ## [3.10.1] - 2023-04-11